# contracts/api-contract-draft.yaml
openapi: 3.1.0
info:
  title: Patika Beosztás API
  version: 0.3.0-phase2b
  description: >-
    A Phase 2B frontend–backend szerződése. A futó backend
    /openapi/v1.json dokumentuma a generált frontend típusok végleges forrása.
    A frontend HTTPS-en, az API-val azonos site alatt fusson, és a cookie-s
    kérésekhez credentials: include beállítást használjon.
servers:
  - url: /
...
paths:
...
  /api/admin/coverage-requirements:
    get:
      tags: [Coverage requirements]
      summary: Lefedettségi szabályok listázása és szűrése
      x-required-permission: ManageCoverageRules
      parameters:
        - in: query
          name: locationId
          schema: { type: [string, 'null'], format: uuid }
        - in: query
          name: dayOfWeek
          schema:
            oneOf:
              - $ref: '#/components/schemas/DayOfWeek'
              - type: 'null'
        - in: query
          name: capability
          schema:
            oneOf:
              - $ref: '#/components/schemas/StaffingCapability'
              - type: 'null'
        - $ref: '#/components/parameters/IncludeInactive'
      responses:
        '200':
          description: Lefedettségi szabályok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CoverageRequirementResponse'
        '404': { $ref: '#/components/responses/Problem' }
    post:
      tags: [Coverage requirements]
      summary: Lefedettségi szabály létrehozása
      x-required-permission: ManageCoverageRules
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCoverageRequirementRequest'
      responses:
        '201':
          description: Létrehozott lefedettségi szabály
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoverageRequirementResponse'
        '404': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /api/admin/coverage-requirements/{id}:
    parameters:
      - $ref: '#/components/parameters/EntityId'
    put:
      tags: [Coverage requirements]
      summary: Lefedettségi szabály módosítása
      x-required-permission: ManageCoverageRules
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCoverageRequirementRequest'
      responses:
        '200':
          description: Módosított lefedettségi szabály
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoverageRequirementResponse'
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /api/admin/coverage-requirements/{id}/deactivate:
    parameters:
      - $ref: '#/components/parameters/EntityId'
    post:
      tags: [Coverage requirements]
      summary: Lefedettségi szabály inaktiválása
      x-required-permission: ManageCoverageRules
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionRequest'
      responses:
        '200':
          description: Inaktivált lefedettségi szabály
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoverageRequirementResponse'
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-PatikaSession
  parameters:
    CsrfToken:
      in: header
      name: X-CSRF-TOKEN
      required: true
      schema:
        type: string
    EntityId:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
...
    IncludeInactive:
      in: query
      name: includeInactive
      schema:
        type: boolean
        default: false
  responses:
    Problem:
      description: RFC 7807 ProblemDetails
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: [string, 'null']
        title:
          type: [string, 'null']
        status:
          type: [integer, 'null']
        detail:
          type: [string, 'null']
        instance:
          type: [string, 'null']
        code:
          type: [string, 'null']
        errors:
          type: [array, 'null']
          items:
            type: object
            properties:
              code: { type: string }
              message: { type: string }
              field: { type: [string, 'null'] }
...
    VersionRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer, minimum: 1 }
...
    DayOfWeek:
      type: string
      enum: [Sunday, Monday, Tuesday, Wednesday, Thursday, Friday, Saturday]
...
    StaffingCapability:
      type: string
      enum: [Pharmacist, SpecialistPharmacist, SpecialistAssistant, Assistant, Cleaner, Finance, Other]
    CoverageSeverity:
      type: string
      enum: [Warning, Blocking]
...
    CreateCoverageRequirementRequest:
      type: object
      additionalProperties: false
      required: [locationId, dayOfWeek, startTime, endTime, requiredCapability, requiredCount, severity, isActive]
      properties:
        locationId: { type: string, format: uuid }
        dayOfWeek: { $ref: '#/components/schemas/DayOfWeek' }
        startTime: { type: string, format: time }
        endTime: { type: string, format: time }
        requiredCapability: { $ref: '#/components/schemas/StaffingCapability' }
        requiredCount: { type: integer, minimum: 1 }
        severity: { $ref: '#/components/schemas/CoverageSeverity' }
        isActive: { type: boolean }
    UpdateCoverageRequirementRequest:
      allOf:
        - $ref: '#/components/schemas/CreateCoverageRequirementRequest'
        - type: object
          required: [expectedVersion]
          properties:
            expectedVersion: { type: integer, minimum: 1 }
    CoverageRequirementResponse:
      allOf:
        - $ref: '#/components/schemas/CreateCoverageRequirementRequest'
        - type: object
          required: [id, locationName, locationIsActive, warnings, version, createdAtUtc, updatedAtUtc]
          properties:
            id: { type: string, format: uuid }
            locationName: { type: string }
            locationIsActive: { type: boolean }
            warnings:
              type: array
              items: { type: string }
            version: { type: integer, minimum: 1 }
            createdAtUtc: { type: string, format: date-time }
            updatedAtUtc: { type: string, format: date-time }
...
security:
  - cookieAuth: []